adios: recompute args when calling an input's functor
Silly bug. When we call a module's functor with some args, we can't get
away with just calling the `impl` of the module. Before, we only
performed a merge with the old args and new args within the functor for
some input module - so you could do `inputs.foo {}`. But a merge isn't
good enough. We have to recompute any/all `defaultFunc`s that the module
might have.